move reused functions into new hashing submodule to avoid import warnings
