+++ spacegraphcats/cdbg/__init__.py
@@ -1,1 +1,1 @@
-from .index_cdbg_by_kmer import hash_sequence, MPHF_KmerIndex  # noqa
+from .hashing import hash_sequence, MPHF_KmerIndex  # noqa
